PERA-1354 Move algod api implementation into foundation
